refactor: change domain into madule
