+++ src/modules/product/entities/products_gen.go
