+++ src/modules/product/repositories/products_repo_query_gen.go
@@ -6,7 +6,7 @@ import (
 	"errors"
 	"fmt"
 	"github.com/jmoiron/sqlx"
-	productsmodel "golang-starter/src/domains/product/entities"
+	productsmodel "golang-starter/src/modules/product/entities"
 	"strings"
 )
