+++ src/modules/user/repositories/users_repo_command_gen.go
@@ -6,7 +6,7 @@ import (
 	"database/sql"
 	"fmt"
 	"github.com/jmoiron/sqlx"
-	usersmodel "golang-starter/src/domains/user/entities"
+	usersmodel "golang-starter/src/modules/user/entities"
 	"strings"
 )
